Initial commit for multinode configuration Mostly worked through prerequisite checking

+++ fmwsetup/scripts/01-undo.sh
@@ -10,11 +10,13 @@
 [[ ! -a setScriptEnv.sh ]] && echo "[> Environment setup could not be completed. Ensure you are executing from the scripts directory, or via the fmw_deploy utility <]" && exit 2 || . setScriptEnv.sh
 [[ $? == "2" ]] && echo "[> Halting script execution <]" && exit 2
 
-echo "> Deleting $FMW_HOME"
-sudo rm -rf $FMW_HOME
-
-echo "> Deleting $DOMAIN_BASE"
-sudo rm -rf $DOMAIN_BASE
+if [[ $MULTINODE == 0 ]]; then
+	echo "> Deleting $FMW_HOME"
+	sudo rm -rf $FMW_HOME
+	
+	echo "> Deleting $DOMAIN_BASE"
+	sudo rm -rf $DOMAIN_BASE
+fi
 
 echo "> Deleting Inventory"
 sudo rm -rf /u01/app/oraInventory
